[Refactor] router to external node module
