+++ app.js
@@ -1,47 +1,6 @@
 const http = require('http');
-const fs = require('fs');
+const routes = require('./routes');
 
-const server = http.createServer((req, res) => {
-  const {url, method} = req;
-
-  if (url === '/') {
-    res.write('<html>');
-    res.write('<head><title>My First Page</title></head>');
-    res.write(`<body>
-      <form action="/message" method="POST">
-        <input type="text" name="message"/>
-        <button type="submit">Send</button>
-      </form>
-    </body>`);
-    res.write('</html>');
-    return res.end();
-  }
-
-  if (url === '/message' && method === 'POST') {
-    const body = [];
-
-    req.on('data', (chunk) => {      
-      body.push(chunk);
-    });
-
-    return req.on('end', () => {
-      const parsedBody = Buffer.concat(body).toString();
-      const [, message] = parsedBody.split("=");
-
-      fs.writeFile('message.txt', message, err => {
-        res.statusCode = 302;
-        res.setHeader('Location', '/');
-        return res.end();
-      });
-    });
-  }
-
-  res.setHeader('Content-Type', 'text/html');
-  res.write('<html>');
-  res.write('<head><title>My First Page</title></head>');
-  res.write('<body><h1>Hello from my Node.js</h1></body>');
-  res.write('</html>');
-  res.end();
-});
+const server = http.createServer(routes);
 
 server.listen(3000);
